add invoices and invoice link

+++ app/Http/Controllers/Backend/Store/StoreController.php
@@ -65,8 +65,17 @@ class StoreController extends Controller{
 
         $name = $store->name;
 
+        if(!$this->deleteable($store)){
+            return redirect()->back()->withErrors('There are Orders in this store, thus can no longer be deleted.');
+        }
+
         $store->delete();
 
         return redirect()->back()->with('success', [$name.' has been deleted.']);
+    }
+
+    protected function deleteable(Store $store)
+    {
+        return $store->orderCount < 1;
     }
 }
